Implemented source generated DummyComAddIn.
